+++ examples/bench.rs
@@ -7,7 +7,7 @@ use eom::*;
 
 fn main() {
     let dt = 0.01;
-    let eom = model::Lorenz63::default();
+    let eom = ode::Lorenz63::default();
     let teo = explicit::rk4(eom, dt);
     let mut buf = teo.new_buffer();
     let mut x: Array1<f64> = arr1(&[1.0, 0.0, 0.0]);
