+++ examples/clv.rs
@@ -58,7 +58,7 @@ pub fn clv<A, S, TEO>(teo: &TEO,
 
 fn main() {
     let dt = 0.01;
-    let eom = model::Lorenz63::default();
+    let eom = ode::Lorenz63::default();
     let teo = explicit::rk4(eom, dt);
     let duration = 100000;
     let ts = clv(&teo, arr1(&[1.0, 0.0, 0.0]), 1e-7, duration);
