+++ examples/goy_shell.rs
@@ -11,7 +11,7 @@ use num_traits::Zero;
 
 fn main() {
     let dt = 1e-5;
-    let eom = model::GoyShell::default();
+    let eom = ode::GoyShell::default();
     let teo = semi_implicit::diag_rk4(eom, dt);
     let mut x0 = rcarr1(&vec![c64::zero(); 27]);
     x0[2] = c64::new(1.0, 0.0);
