+++ tests/view.rs
@@ -8,7 +8,7 @@ use eom::*;
 #[test]
 fn arr() {
     let dt = 0.01;
-    let eom = model::Lorenz63::default();
+    let eom = ode::Lorenz63::default();
     let teo = explicit::euler(eom, dt);
     let mut x: Array1<f64> = arr1(&[1.0, 0.0, 0.0]);
     let mut buf = teo.new_buffer();
@@ -18,7 +18,7 @@ fn arr() {
 #[test]
 fn rcarr() {
     let dt = 0.01;
-    let eom = model::Lorenz63::default();
+    let eom = ode::Lorenz63::default();
     let teo = explicit::euler(eom, dt);
     let mut x: RcArray1<f64> = rcarr1(&[1.0, 0.0, 0.0]);
     let mut buf = teo.new_buffer();
@@ -28,7 +28,7 @@ fn rcarr() {
 #[test]
 fn view_mut() {
     let dt = 0.01;
-    let eom = model::Lorenz63::default();
+    let eom = ode::Lorenz63::default();
     let teo = explicit::euler(eom, dt);
     let mut x: Array1<f64> = arr1(&[1.0, 0.0, 0.0]);
     let mut v = &mut x.view_mut();
